# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Waffle Hub BCM - KOT & Billing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  displayOrder: number;
  active: boolean;
  createdAt: string;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  displayOrder?: number;
  active?: boolean;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  displayOrder?: number;
  active?: boolean;
}

export interface Product {
  id: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  active: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
  active?: boolean;
}

export interface ProductUpdate {
  /** @nullable */
  categoryId?: number | null;
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price?: number;
  active?: boolean;
}

export interface OrderItem {
  id: number;
  orderId: number;
  /** @nullable */
  productId?: number | null;
  productName: string;
  price: number;
  quantity: number;
  /** @nullable */
  notes?: string | null;
}

export interface Payment {
  id: number;
  orderId: number;
  totalAmount: number;
  cashAmount: number;
  upiAmount: number;
  cardAmount: number;
  totalPaid: number;
  balance?: number;
  status: string;
  createdAt: string;
}

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  orderType: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  totalAmount: number;
  createdAt: string;
  readyTime: string;
  updatedAt?: string;
  items: OrderItem[];
  payment?: Payment;
}

export interface OrderItemInput {
  /** @nullable */
  productId?: number | null;
  /** @minLength 1 */
  productName: string;
  price: number;
  /** @minimum 1 */
  quantity: number;
  notes?: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  customerPhone?: string;
  /** @nullable */
  customerId?: number | null;
  orderType: string;
  notes?: string;
  items?: OrderItemInput[];
}

export interface OrderUpdate {
  customerName?: string;
  customerPhone?: string;
  orderType?: string;
  /** @nullable */
  notes?: string | null;
}

export interface OrderStatusUpdate {
  status: string;
}

export interface OrderItemUpdate {
  /** @minimum 1 */
  quantity?: number;
  /** @nullable */
  notes?: string | null;
}

export interface PaymentInput {
  totalAmount: number;
  cashAmount?: number;
  upiAmount?: number;
  cardAmount?: number;
}

export interface PaymentUpdate {
  cashAmount?: number;
  upiAmount?: number;
  cardAmount?: number;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  totalSpending: number;
  orderCount: number;
  /** @nullable */
  lastOrderDate?: string | null;
  /** @nullable */
  favoriteItems?: string | null;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
}

export interface CustomerUpdate {
  name?: string;
  phone?: string;
}

export interface Expense {
  id: number;
  title: string;
  amount: number;
  paymentMethod: string;
  /** @nullable */
  notes?: string | null;
  expenseDate?: string;
  createdAt: string;
}

export interface ExpenseInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 0 */
  amount: number;
  paymentMethod: string;
  notes?: string;
  expenseDate?: string;
}

export interface ExpenseUpdate {
  title?: string;
  amount?: number;
  paymentMethod?: string;
  /** @nullable */
  notes?: string | null;
  expenseDate?: string;
}

export interface DashboardSummary {
  pendingCount: number;
  preparingCount: number;
  readyCount: number;
  completedToday: number;
  cancelledToday: number;
  todayRevenue: number;
  todayOrders: number;
  todayExpenses: number;
  recentOrders?: Order[];
}

export interface DailyReport {
  date: string;
  totalOrders: number;
  totalRevenue: number;
  cashRevenue: number;
  upiRevenue: number;
  cardRevenue: number;
  totalExpenses: number;
  estimatedProfit: number;
  avgOrderValue: number;
  completedOrders: number;
  cancelledOrders: number;
}

export interface ProductSalesItem {
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export type ListProductsParams = {
categoryId?: number;
active?: boolean;
};

export type ListOrdersParams = {
status?: string;
search?: string;
date?: string;
};

export type ListCustomersParams = {
search?: string;
};

export type ListExpensesParams = {
period?: string;
date?: string;
};

export type GetDailyReportParams = {
date?: string;
};

export type GetProductReportParams = {
period?: string;
};