# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Waffle Hub BCM - KOT & Billing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Products ─────────────────────────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── Expenses ─────────────────────────────────────────────────────────────────
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Dashboard & Reports ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, displayOrder, active, createdAt]
      properties:
        id:           { type: integer }
        name:         { type: string }
        displayOrder: { type: integer }
        active:       { type: boolean }
        createdAt:    { type: string }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:         { type: string, minLength: 1 }
        displayOrder: { type: integer }
        active:       { type: boolean }
    CategoryUpdate:
      type: object
      properties:
        name:         { type: string, minLength: 1 }
        displayOrder: { type: integer }
        active:       { type: boolean }

    # ── Product ───────────────────────────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, name, phone, totalSpending, orderCount, createdAt]
      properties:
        id:            { type: integer }
        name:          { type: string }
        phone:         { type: string }
        totalSpending: { type: number }
        orderCount:    { type: integer }
        lastOrderDate: { type: ["string", "null"] }
        favoriteItems: { type: ["string", "null"] }
        createdAt:     { type: string }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name:  { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
    CustomerUpdate:
      type: object
      properties:
        name:  { type: string }
        phone: { type: string }

    # ── Expense ───────────────────────────────────────────────────────────────
    Expense:
      type: object
      required: [id, title, amount, paymentMethod, createdAt]
      properties:
        id:            { type: integer }
        title:         { type: string }
        amount:        { type: number }
        paymentMethod: { type: string }
        notes:         { type: ["string", "null"] }
        expenseDate:   { type: string }
        createdAt:     { type: string }
    ExpenseInput:
      type: object
      required: [title, amount, paymentMethod]
      properties:
        title:         { type: string, minLength: 1 }
        amount:        { type: number, minimum: 0 }
        paymentMethod: { type: string }
        notes:         { type: string }
        expenseDate:   { type: string }
    ExpenseUpdate:
      type: object
      properties:
        title:         { type: string }
        amount:        { type: number }
        paymentMethod: { type: string }
        notes:         { type: ["string", "null"] }
        expenseDate:   { type: string }

    # ── Dashboard ─────────────────────────────────────────────────────────────
...