# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Waffle Hub BCM - KOT & Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  displayOrder?: number;
  active?: boolean;
}