# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Waffle Hub BCM - KOT & Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  phone?: string;
}